organize ChaosIndicatorsViewModel into MVVM
