+++ Earth_Final_Season/Earth_Final_Season/Earth_Final_SeasonApp.swift
@@ -10,11 +10,19 @@ import SwiftUI
 @main
 struct Earth_Final_SeasonApp: App {
     @State private var gameplayVM = GameplayViewModel()
+    @State private var chaosIndicatorsVM: ChaosIndicatorsViewModel
+
+    init() {
+        let gameplayViewModel = GameplayViewModel()
+        _chaosIndicatorsVM = State(wrappedValue: ChaosIndicatorsViewModel(gameplayVM: gameplayViewModel))
+        _gameplayVM = State(wrappedValue: gameplayViewModel)
+    }
     
     var body: some Scene {
         WindowGroup {
             MenuView()
                 .environment(gameplayVM)
+                .environment(chaosIndicatorsVM)
             .task {
                 SoundtrackAudioManager.shared.playSound(named: "lowtoneST")
             }
